Moved atomics to runtime/atomics
